refactor: rearrange project
