+++ src/cli/mod.rs
@@ -1,0 +1,2 @@
+pub mod app;
+pub mod handlers;
